rename fixtures.py to factories.py
